backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
google-generativeai==0.7.2
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==42.0.5
PyPDF2==3.0.1
python-docx==1.1.0
werkzeug==3.1.6
